refactor(stm): relocate circuit verification key provider factories
